Update postgres client
There doesn't seem to be the need to use the 3.7 repo as the client keeps
changing anyway. We've been ok using the client from the edge repo with
Heroku over the last few months. By using the default postgres client we
do not have to worry about updating the postgres client version when it
has been updated/no longer available.